refactor: 인수테스트 메소드 분리
