+++ src/test/java/nextstep/subway/path/acceptance/PathAcceptanceTest.java
@@ -2,25 +2,21 @@ package nextstep.subway.path.acceptance;
 
 import static nextstep.subway.line.acceptance.LineAcceptanceTestMethod.지하철_노선_등록되어_있음;
 import static nextstep.subway.line.acceptance.LineSectionAcceptanceTestMethod.지하철_노선에_지하철역_등록_요청;
-import static org.assertj.core.api.Assertions.assertThat;
+import static nextstep.subway.path.acceptance.PathAcceptanceTestMethod.지하철_최단_경로_조회_요청_응답됨;
+import static nextstep.subway.path.acceptance.PathAcceptanceTestMethod.지하철_최단_경로_조회됨;
+import static nextstep.subway.path.acceptance.PathAcceptanceTestMethod.최단_경로_조회_요청;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 
 @DisplayName("지하철 경로 조회")
@@ -80,36 +76,6 @@ public class PathAcceptanceTest extends AcceptanceTest {
 
         //then : 최단 거리가 조회된다.
         지하철_최단_경로_조회됨(최단_경로_조회_요청, Arrays.asList(교대역, 남부터미널역, 양재역), 5);
-    }
-
-    public static ExtractableResponse<Response> 최단_경로_조회_요청(Long sourceStationId, Long targetStationId) {
-        return RestAssured.given().log().all()
-                .when()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .param("sourceId", sourceStationId)
-                .param("targetId", targetStationId)
-                .get("/paths")
-                .then().log().all()
-                .extract();
-    }
-
-    public static void 지하철_최단_경로_조회됨(ExtractableResponse<Response> response, List<StationResponse> expectedStations,
-                                     int distance) {
-        PathResponse path = response.as(PathResponse.class);
-        List<Long> stationIds = path.getStations().stream()
-                .map(it -> it.getId())
-                .collect(Collectors.toList());
-
-        List<Long> expectedStationIds = expectedStations.stream()
-                .map(it -> it.getId())
-                .collect(Collectors.toList());
-
-        assertThat(stationIds).containsExactlyElementsOf(expectedStationIds);
-        assertThat(path.getDistance()).isEqualTo(distance);
-    }
-
-    public static void 지하철_최단_경로_조회_요청_응답됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
     }
